fix: update PUT body to match forem's API
The PUT API for updating an article expects an `article` property
on the body that that `body_markdown` is a part of. This should be
a pretty simple API change along with the interface update.

As a side note, I was not able to get tests running locally,
but I'll adjust accordingly as I need to from the CI  results.

closes #35